Add support for large folders (each track variable covers 2 bytes now)*. Add support for advertisement folder (100). Adjust RFID tool accordingly. Increase large change step to 50 for options > 255.
*: - Folder 1 to 15 can have up to 3000 tracks
    - Each track > 255 must be named with 4 digits at the start: <0256_myFile.mp3> up to <3000_myLastFile.mp3>
    - All tracks from 1 to 255 stay at 3 digit format to support backward compatibility and be able to move those files to non-large folders without renaming
**: - Advertisement music cards with single mode can be used for memory (listen to songs while playing memory with advertisement tracks)

+++ Tonuino_Main/src/shared/Tonuino_EEPROM.h
@@ -11,7 +11,7 @@ class TonuinoEEPROM
 	void resetEEPROM();
 	
 	void writeToFlash_RecentMusicDS(MusicDataset dataset);
-	void writeToFlash_RecentTrack(uint8_t track);
+	void writeToFlash_RecentTrack(uint16_t track);
 	void writeToFlash_RecentFolder(uint8_t folder);
 	
 	MusicDataset loadFromFlash_RecentMusicDS();
